+++ app/src/main/AndroidManifest.xml
@@ -86,6 +86,9 @@
             android:exported="false"
             android:theme="@style/SlideAnimation" />
         <activity
+            android:name=".presentation.ui.diary.DiaryActivity"
+            android:exported="false" />
+        <activity
             android:name=".presentation.ui.group.GroupInfoActivity"
             android:exported="false"
             android:theme="@style/Theme.MaterialTransparent"
