+++ app/src/main/java/com/mongmong/namo/presentation/config/FilterState.kt
@@ -1,0 +1,8 @@
+package com.mongmong.namo.presentation.config
+
+enum class FilterState(val state: String) {
+    NONE("없음"),
+    TITLES("일정 제목"),
+    CONTENTS("기록 내용"),
+    PARTICIPANTS("참석자")
+}
