+++ app/src/main/res/drawable/bg_filter_btn.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <corners android:radius="4dp"/>
+    <solid
+        android:color="@color/main"/>
+</shape>
