+++ app/src/main/res/layout/activity_diary.xml
@@ -1,0 +1,140 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <data>
+        <import type="android.view.View"/>
+        <variable
+            name="viewModel"
+            type="com.mongmong.namo.presentation.ui.diary.DiaryViewModel" />
+    </data>
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/framelayout"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+        <!-- 툴바 -->
+        <LinearLayout
+            android:id="@+id/diary_month_ll"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginHorizontal="24dp"
+            android:gravity="center_vertical"
+            app:layout_constraintTop_toTopOf="parent">
+            <ImageButton
+                android:id="@+id/diary_back_btn"
+                android:layout_width="32dp"
+                android:layout_height="32dp"
+                android:layout_marginVertical="10dp"
+                android:background="@color/transparent"
+                android:src="@drawable/ic_back" />
+            <TextView
+                style="@style/header"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/diary_storage"
+                android:fontFamily="@font/pretendard_bold"
+                android:textColor="@color/black"
+                android:textSize="16sp"
+                android:layout_weight="1"
+                android:gravity="center" />
+
+            <ImageButton
+                android:id="@+id/diary_calendar_btn"
+                android:layout_width="32dp"
+                android:layout_height="32dp"
+                android:layout_marginVertical="10dp"
+                android:background="@color/transparent"
+                android:src="@drawable/ic_calendar" />
+        </LinearLayout>
+
+        <TextView
+            style="@style/content_regular"
+            android:id="@+id/diary_filter"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="30dp"
+            android:background="@drawable/bg_diary_tab"
+            android:text="@{String.valueOf(viewModel.filter)}"
+            android:drawableLeft="@drawable/ic_filter"
+            android:drawablePadding="8dp"
+            android:drawableTint="@{viewModel.filter.NONE ? @color/main_gray : @color/main}"
+            android:paddingHorizontal="12dp"
+            android:paddingVertical="6dp"
+            app:layout_constraintTop_toBottomOf="@id/diary_month_ll"
+            app:layout_constraintStart_toStartOf="parent"/>
+
+        <EditText
+            style="@style/filter_edit_text"
+            android:id="@+id/diary_filter_keyword_et"
+            android:hint="@string/diary_filter_keyword_hint"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="30dp"
+            android:layout_marginEnd="32dp"
+            android:layout_marginTop="8dp"
+            android:text="@={viewModel.keyword}"
+            app:layout_constraintEnd_toStartOf="@+id/diary_filter_search_btn"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/diary_filter" />
+
+        <androidx.appcompat.widget.AppCompatButton
+            style="@style/filter_btn"
+            android:id="@+id/diary_filter_search_btn"
+            android:layout_width="58dp"
+            android:layout_height="30dp"
+            android:layout_marginEnd="30dp"
+            android:text="@string/diary_filter_search"
+            app:layout_constraintBottom_toBottomOf="@id/diary_filter_keyword_et"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="@id/diary_filter_keyword_et" />
+
+        <LinearLayout
+            android:id="@+id/diary_list_empty_layout"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintTop_toBottomOf="@id/diary_month_ll"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:layout_marginTop="160dp"
+            android:orientation="vertical"
+            android:visibility="@{viewModel.isListEmpty ? View.VISIBLE : View.GONE}">
+            <ImageView
+                android:id="@+id/diary_list_empty_iv"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:src="@drawable/img_diary_empty"
+                android:imageSource="@{viewModel.emptyImageResId}"
+                android:layout_gravity="center"/>
+            <TextView
+                style="@style/content_regular"
+                android:id="@+id/diary_list_empty_tv"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="45dp"
+                android:text="@{viewModel.emptyMessageResId}"
+                android:textAlignment="center"
+                android:layout_gravity="center" />
+        </LinearLayout>
+
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/diary_list_rv"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:background="@color/white"
+            android:clipToPadding="false"
+            android:orientation="vertical"
+            android:paddingBottom="16dp"
+            android:layout_marginTop="20dp"
+            android:visibility="@{!viewModel.isListEmpty ? View.VISIBLE : View.GONE}"
+            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/diary_filter_keyword_et"
+            tools:listitem="@layout/item_diary_item_list" />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
