--- app/src/main/res/layout/fragment_diary.xml
@@ -1,129 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-        <import type="android.view.View"/>
-        <variable
-            name="viewModel"
-            type="com.mongmong.namo.presentation.ui.diary.DiaryViewModel" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/framelayout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <LinearLayout
-            android:id="@+id/diary_month_ll"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="20dp"
-            android:layout_marginTop="20dp"
-            android:gravity="center_vertical"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
-            <TextView
-                android:id="@+id/diary_month"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@{viewModel.currentDate}"
-                tools:text="yyyy년 MM월"
-                style="@style/header"/>
-
-            <androidx.appcompat.widget.AppCompatImageView
-                android:id="@+id/diary_month_arrow_iv"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="13dp"
-                android:src="@drawable/ic_arrow_down_bold"
-                app:tint="@color/black"
-                app:layout_constraintTop_toTopOf="@id/diary_month"
-                app:layout_constraintBottom_toBottomOf="@id/diary_month"
-                app:layout_constraintStart_toEndOf="@id/diary_month"/>
-        </LinearLayout>
-
-        <com.google.android.material.tabs.TabLayout
-            android:id="@+id/diary_tab"
-            android:layout_width="152dp"
-            android:layout_height="30dp"
-            android:layout_marginEnd="20dp"
-            android:background="@drawable/bg_diary_tab"
-            app:tabIndicator="@drawable/bg_diary_tab_indicator"
-            app:tabIndicatorColor="@color/main"
-            app:tabIndicatorHeight="30dp"
-            app:tabMode="auto"
-            app:tabIndicatorAnimationMode="elastic"
-            app:tabTextAppearance="@style/TabTextStyle"
-            app:tabTextColor="@color/main_text"
-            app:tabSelectedTextColor="@color/white"
-            app:tabRippleColor="@null"
-            app:layout_constraintTop_toTopOf="@id/diary_month_ll"
-            app:layout_constraintBottom_toBottomOf="@id/diary_month_ll"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:onTabSelectedListener="@{viewModel.onTabSelectedListener}"/>
-
-        <LinearLayout
-            android:id="@+id/diary_list_empty_layout"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:layout_constraintTop_toBottomOf="@id/diary_month_ll"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            android:layout_marginTop="160dp"
-            android:orientation="vertical"
-            android:visibility="@{viewModel.isListEmpty ? View.VISIBLE : View.GONE}">
-            <ImageView
-                android:id="@+id/diary_list_empty_iv"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:src="@drawable/img_diary_empty"
-                android:imageSource="@{viewModel.emptyImageResId}"
-                android:layout_gravity="center"/>
-            <TextView
-                android:id="@+id/diary_list_empty_tv"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="45dp"
-                android:text="@{viewModel.emptyMessageResId}"
-                android:textAlignment="center"
-                android:layout_gravity="center" />
-        </LinearLayout>
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/diary_personal_list_rv"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_marginTop="10dp"
-            android:paddingBottom="16dp"
-            android:clipToPadding="false"
-            android:background="@color/white"
-            android:orientation="vertical"
-            android:visibility="@{viewModel.isMoim == viewModel.IS_NOT_MOIM &amp;&amp; !viewModel.isListEmpty ? View.VISIBLE : View.GONE}"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/diary_month_ll"
-            tools:listitem="@layout/item_diary_list" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/diary_group_list_rv"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_marginTop="10dp"
-            android:paddingBottom="16dp"
-            android:clipToPadding="false"
-            android:background="@color/white"
-            android:orientation="vertical"
-            android:visibility="@{viewModel.isMoim == viewModel.IS_MOIM &amp;&amp; !viewModel.isListEmpty ? View.VISIBLE : View.GONE}"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/diary_month_ll"
-            tools:listitem="@layout/item_diary_item_list" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
