+++ app/src/main/res/layout/fragment_home.xml
@@ -42,11 +42,20 @@
                     android:layout_marginStart="13dp"
                     android:src="@drawable/ic_arrow_down_bold"
                     app:tint="@color/black"
-                    app:layout_constraintTop_toTopOf="@id/diary_month"
-                    app:layout_constraintBottom_toBottomOf="@id/diary_month"
-                    app:layout_constraintStart_toEndOf="@id/diary_month"/>
+                    app:layout_constraintTop_toTopOf="@id/home_calendar_year_month_tv"
+                    app:layout_constraintBottom_toBottomOf="@id/home_calendar_year_month_tv"
+                    app:layout_constraintStart_toEndOf="@id/home_calendar_year_month_tv"/>
             </LinearLayout>
 
+            <ImageButton
+                android:id="@+id/home_calendar_diary_collect_btn"
+                android:layout_width="24dp"
+                android:layout_height="24dp"
+                android:src="@drawable/ic_diary_stroage"
+                android:background="@color/transparent"
+                android:layout_marginEnd="20dp"
+                app:layout_constraintEnd_toStartOf="@+id/home_calendar_today_tv"
+                app:layout_constraintTop_toTopOf="@+id/home_calendar_today_tv" />
             <TextView
                 android:id="@+id/home_calendar_today_tv"
                 android:layout_width="25dp"
