--- app/src/main/res/values-v26/styles.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <style name="map_common_btn">
-        <item name="android:layout_width">140dp</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:paddingVertical">6dp</item>
-        <item name="android:background">@drawable/bg_map_btn</item>
-        <item name="android:gravity">center</item>
-        <item name="android:textColor">@color/main</item>
-        <item name="android:textSize">15sp</item>
-        <item name="android:textStyle">bold</item>
-    </style>
-</resources>
