+++ app/src/main/res/values/strings.xml
@@ -65,6 +65,10 @@
     <string name="schedule_moim_edit">모임 일정 편집</string>
 
     <!--diary 관련 String-->
+    <string name="diary_storage">보관함</string>
+    <string name="diary_filter">검색 필터 없음</string>
+    <string name="diary_filter_search">검색</string>
+    <string name="diary_filter_keyword_hint">키워드 입력</string>
     <string name="diary_network_failure">모임 기록은 네트워크 연결 시 확인할 수 있습니다.\n네트워크 연결 후 다시 시도해 주세요.</string>
     <string name="diary_empty">기록이 없습니다.\n일정에 기록을 남겨보세요!</string>
     <string name="diary_modify">수정</string>
