+++ app/src/main/res/values/styles.xml
@@ -185,6 +185,25 @@
         <item name="android:textAlignment">center</item>
     </style>
 
+    <!-- 필터 -->
+    <style name="filter_edit_text">
+        <item name="android:textColor">@color/main_text</item>
+        <item name="android:textSize">15sp</item>
+        <item name="android:textColorHint">@color/text_placeholder</item>
+        <item name="android:backgroundTint">@color/text_placeholder</item>
+        <item name="android:maxLines">1</item>
+        <item name="android:drawableLeft">@drawable/ic_search</item>
+        <item name="android:drawablePadding">8dp</item>
+        <item name="android:drawableTint">@color/text_placeholder</item>
+    </style>
+    <style name="filter_btn">
+        <item name="android:background">@drawable/bg_filter_btn</item>
+        <item name="android:gravity">center</item>
+        <item name="android:textColor">@color/white</item>
+        <item name="android:textSize">15sp</item>
+        <item name="android:fontFamily">@font/pretendard_bold</item>
+        <item name="elevation">0dp</item>
+    </style>
     <!-- 일정 -->
     <style name="empty_text">
         <item name="android:layout_width">wrap_content</item>
